Refactor manifest as single class
