+++ packages/11ty/_plugins/iiif/manifest/writer.js
