--- packages/11ty/_plugins/iiif/process/manifest/base.js
@@ -1,9 +1,0 @@
-module.exports = class Base {
-  constructor(data) {
-    this.iiifConfig = data
-  }
-
-  getBaseId(figure) {
-    return [process.env.URL, this.iiifConfig.outputDir, figure.id].join('/')
-  }
-}
